# openapi/poshub_v1.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "PosHub API",
    "description": "# PosHub API Documentation\n\nA comprehensive Point of Sale (POS) system API built with FastAPI.\n\n## Features\n\n- 🔐 **JWT Authentication**: Secure token-based authentication with scope-based permissions\n- 📦 **Order Management**: Complete order lifecycle management\n- 🌐 **External Integrations**: Safe HTTP client for external API calls\n- 🏥 **Health Monitoring**: System health checks and monitoring\n- 📊 **OpenAPI Export**: Full OpenAPI 3.0 schema export capabilities\n\n## Authentication\n\nThis API uses JWT tokens for authentication. Tokens must be included in the Authorization header:\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\n## Scopes\n\nThe API uses the following scopes for authorization:\n- `orders:read`: Read access to orders\n- `orders:write`: Write access to orders (includes read)\n\n## Rate Limiting\n\nAPI calls are subject to rate limiting. Please refer to the response headers for current limits.\n\n## Error Handling\n\nAll API errors follow RFC 7807 Problem Details format with correlation IDs for tracing.",
    "version": "1.0.0",
    "contact": {
      "name": "PosHub API Support",
      "email": "support@poshub.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000/v1",
...
    }
...
  "paths": {
...
    "/v1/orders": {
      "post": {
        "summary": "Create Order",
        "description": "Create a new order in the system",
        "operationId": "create_order_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderIn" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderOut" }
              }
            }
          },
          "400": {
            "description": "Invalid order data",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid order data provided"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticated"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions. Required scope: orders:write"
                }
              }
            }
          },
          "409": {
            "description": "Order already exists",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Order with this ID already exists"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": ["body", "customer_id"],
                      "msg": "field required",
                      "type": "value_error.missing"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["orders"],
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/v1/orders/{order_id}": {
      "get": {
        "summary": "Get Order",
        "description": "Retrieve an order by its unique ID",
        "operationId": "get_order_v1_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderOut" }
              }
            }
          },
          "400": {
            "description": "Invalid order ID format",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid UUID format"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Not authenticated"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Insufficient permissions. Required scope: orders:read"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Order not found"
                }
              }
            }
          }
        },
        "tags": ["orders"],
        "security": [{ "HTTPBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderIn": {
        "type": "object",
        "properties": {
          "nom_client": {
            "type": "string",
            "maxLength": 128,
            "description": "Customer name"
          },
          "montant": {
            "type": "number",
            "minimum": 0,
            "description": "Total amount"
          },
          "devise": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "description": "Currency code (ISO 4217)"
          }
        },
        "additionalProperties": false,
        "required": ["nom_client", "montant", "devise"],
        "title": "OrderIn",
        "description": "Order input model for creating new orders",
        "example": {
          "nom_client": "hasna",
          "montant": 99.99,
          "devise": "EUR"
        }
      },
      "OrderOut": {
        "type": "object",
        "properties": {
          "order": {
            "type": "string",
            "format": "uuid",
            "description": "Unique order identifier"
          },
          "nom_client": {
            "type": "string",
            "description": "Customer name"
          },
          "montant": {
            "type": "number",
            "description": "Total amount"
          },
          "devise": {
            "type": "string",
            "description": "Currency code"
          },
          "created_by": {
            "type": "string",
            "description": "User who created the order",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "required": ["order", "nom_client", "montant", "devise"],
        "title": "OrderOut",
        "description": "Order output model with system-generated fields",
        "example": {
          "order": "123e4567-e89b-12d3-a456-426614174000",
          "nom_client": "hasna",
          "montant": 99.99,
          "devise": "EUR",
          "created_by": "admin"
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      }
    }
  },
...
} 